Reset modem if carrier lost (#94)
When carrier is lost the PLL keeps looking for it and starts drifting
slowly. This can make it harder to lock on to the carrier when it
comes back. One solution is to reset the frequency back to 57 Khz
after 10 seconds of loss of block sync being detected.

Open question: is the PLL interface being used correctly?